"the no" - big bee

+++ app.py
@@ -216,6 +216,8 @@ def search(*query:str): # Search for pkgs
 def init(*, version:str=None, modloader:str=None, quiet:bool=False, override:bool=False): # Init project
     global kjspkgfile
 
+    if not _check_project(): _err("Hmm... This directory doesn't look like a kubejs directory") # Wrong dir err
+
     if _project_exists(): # Override
         if not quiet and input("\u001b[31;1mA PROJECT ALREADY EXISTS IN THIS REPOSITORY, CREATING A NEW ONE OVERRIDES THE PREVIOUS ONE, ARE YOU SURE YOU WANT TO PROCEED? (y/N): \u001b[0m").lower()=="y" or override: _delete_project()
         else: exit(0)
@@ -289,18 +291,19 @@ def _parser(func:str="help", *args, help:bool=False, **kwargs):
 
     if func not in FUNCTIONS.keys(): _err("Command \""+func+"\" is not found. Run \"kjspkg help\" to see all of the available commands") # Wrong command err
     
+    helperfunc = False
     if FUNCTIONS[func] not in (info, init, pkginfo, search): # If the command is not a any-dir command
-        if not _check_project(): _err("Hmm... This directory doesn't look like a kubejs directory") # Wrong dir err
         if not _project_exists(): # If a project is not found, call init
             print(_bold("Project not found, a new one will be created.\n"))
             init()
+        if path.exists(".kjspkg"): kjspkgfile = load(open(".kjspkg")) # Open .kjspkg
 
-    if path.exists(".kjspkg"): kjspkgfile = load(open(".kjspkg")) # Open .kjspkg
+        helperfunc = True # Set the helperfunc var
 
     FUNCTIONS[func](*args, **kwargs) # Run the command
 
     # Clean up
-    if path.exists(".kjspkg"): # If uninit wasn't called
+    if path.exists(".kjspkg") and not helperfunc: # If uninit wasn't called and the command isn't a help command
         with open(".kjspkg", "w") as f: dump(kjspkgfile, f) # Save .kjspkg
 
 # RUN
